frontend: make navbar resizable
